Update 3rd party find logic for _ROOT support
Ideally, we should just be able to pass *_ROOT variables on the command
line to CMake subbuilds and have find_package do the right thing.  For
zlib that means the 3rd party ZLIB modules need to know about the z_brl
convention, and for SQLite3 we need to update CMake's own modern
FindSQLite3.cmake to support setting a root value (and update proj to
use the newer variable names from upstream.)

The goal is to avoid the messy business of figuring out what exact
filename to spoof the _LIBRARY variables with per platform.  The
tradeoff, if we can't upstream these pieces, will be having to mod the
sources so *_ROOT works...  probably worthwhile overall, but time will
tell.