Reorganize so each project has a toplevel directory
Per an earlier discussion about what this repository should look like,
reorganize so each project has a "parent" directory with the actual
source code in a subdirectory.  The idea will be that patch files and
other resources we need to maintain beyond the vanilla source trees will
be located in these subdirectories.