Checkpoint some experiments with Windows Qt build
Yikes.  Strange symptoms with the Qt ExternalProject_Add where the
configure would work but not generate its stamp file properly, and then
the build/install steps would not do anything (work or stamping) but
also report success.

Wrapping the invocations in CMake scripts seems to at least allow the
build/install to complete, although right now it seems to be producing a
debug build based on the filenames...