Remove prefixes for OSX building
The OSX binaries seem to be a bit more sensitive to overwriting these
paths - just don't put the prefixes in to begin with.  If non-full
paths were going to break anything we'd be facing that problem whether
we eliminate them after the fact or before building... the drawback of
this is that it requires patching rather than just post processing the
binary outputs generically and directly.  However, if OSX won't tolerate
the latter in the case of the package path we'll have to go this route.