Manage non-CMake Qt build configure with scripts
For multi-config CMake generators, we want the Qt configure to know if
we are doing Debug or Release building so it can pass the appropriate
flag to configure/configure.bat - however, at CMake configure time we
don't know what the user will build for multiconfig generators and
there's no handy CMake generator expression for Qt configure options.

Instead, we set up scripts that are passed the output of generator
expressions at build time.  The scripts then do the translations to get
the correct configure arguments.  While we're at it, consolidate all the
platforms into three scripts, one for each stage.