See if we can make utahrle into a brlcad_externals repo
In principle, I don't know of any reason we couldn't use a system
version of this, if we ever found one (unlike openNURBS).

If we want to do this for openNURBS as well, we'll have to do not just a
find_package check but also some functionality tests to look for our
mods being present to accept or reject a "system" openNURBS (which will
probably always end up rejecting...)  On the other hand, the patch
approach to openNURBS would be beneficial in that it would keep clear
what our changes are, so it might be worth considering anyway.