Check for opennurbs_x.h when doing find_package
There are a number openNURBS changes we should be checking for to
properly validate whether an openNURBS install is appropriate for
BRL-CAD, but for now we'll just do the simple one and look for the
opennurbs_x.h header.  If that's not present the install is no good.

Realistically a system install of openNURBS, even if we should actually
find one somewhere, is almost certainly not going to be what BRL-CAD
needs.  However, going the patch + find_package route with it may, in
principle, make updating to future openNURBS releases simpler.  Also,
the requirement for plief and strclear when doing BRL-CAD builds means
that the BRL-CAD build will, practically speaking, always require a
brlcad_externals output directory anyway.  Since that's the case, we may
as well "promote" things in src/other with active upstreams to
brlcad_externals to try and make syncing simpler.