Try capturing Qt output to a variable.
We'll probably want to write this to log files, but Chris dug into the
Windows build and the output from the Qt build process seems to be
interfering with the Visual Studio build environment in some fashion.
See if we can silence the output as far as Visual Studio is concerned by
capturing to variables in CMake.