Got a running OSPRay ospExamples on Ubuntu
Using system m4 and python3, and undoubtedly need a lot more tweaks to
the find logic in various packages to use bundled zlib/png/etc. rather
than spotting system versions, but this gets us as far as a running
OSPRay example program on Ubuntu Linux.

Probably will stop here for now with OSPRay, since the current focus is
Appleseed, but this gives us a good idea of what supporting OSPRay will
look like if and when.

M4 is likely the most problematic of the ISPC dependencies to get set up
as a readily buildable dependency, but Python may be more challenging
than I'd like as well.  I hit what appears to be this issue on Linux:
https://github.com/python/cpython/issues/90737 and I'm not even sure yet
how to proceed with a build and install on Windows.